@@ -1,3 +1,7 @@
 Rails.application.routes.draw do
-  resources :kittens
+  scope '', as: 'plugin_kitten' do
+    scope 'projects/:project_id', as: 'project' do
+      resources :kittens
+    end
+  end
 end
